[Refactor] : presentation 모듈 그룹화
Git : #6
