+++ lib/app/main.dart
@@ -1,4 +1,4 @@
-import 'package:earlybuddy/presentation/addschedule/example/addschedule_example.dart';
+import 'package:earlybuddy/presentation/schedule/addschedule/example/addschedule_example.dart';
 
 import 'package:flutter/material.dart';
 import 'package:intl/date_symbol_data_local.dart';
