+++ lib/presentation/auth/auth_model/email.dart
