+++ lib/presentation/auth/auth_model/model.dart
