+++ lib/presentation/auth/login/view/login_view.dart
@@ -1,6 +1,6 @@
 import 'package:earlybuddy/domain/auth/auth_repository.dart';
-import 'package:earlybuddy/presentation/login/bloc/login_bloc.dart';
-import 'package:earlybuddy/presentation/register/register.dart';
+import 'package:earlybuddy/presentation/auth/login/bloc/login_bloc.dart';
+import 'package:earlybuddy/presentation/auth/register/register.dart';
 import 'package:earlybuddy/shared/eb_resources/eb_resources.dart';
 import 'package:earlybuddy/shared/eb_uikit/eb_uikit.dart';
 import 'package:flutter/material.dart';
