+++ lib/presentation/auth/register/bloc/register_state.dart
