+++ lib/presentation/auth/register/view/appbar.dart
