+++ lib/presentation/schedule/map/bloc/map_bloc.dart
@@ -1,0 +1,2 @@
+part 'map_event.dart';
+part 'map_state.dart';
