+++ lib/presentation/schedule/map/bloc/map_event.dart
@@ -1,0 +1,1 @@
+part of 'map_bloc.dart';
