+++ lib/presentation/schedule/map/bloc/map_state.dart
@@ -1,0 +1,1 @@
+part of 'map_bloc.dart';
