+++ lib/presentation/schedule/map/map.dart
@@ -1,0 +1,1 @@
+export 'view/view.dart';
