+++ lib/presentation/schedule/map/view/map_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class MapView extends StatelessWidget {
+  const MapView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(body: Text('MapView'));
+  }
+}
