+++ lib/presentation/schedule/map/view/view.dart
@@ -1,0 +1,1 @@
+export 'map_view.dart';
