+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   freezed: ^2.5.1
   freezed_annotation: ^2.4.1
   json_annotation: ^4.8.1
+  modal_bottom_sheet: ^3.0.0
 
 dev_dependencies:
   flutter_test:
